+++ tests/test_schema.py
@@ -7,6 +7,7 @@ from threedi_modelchecker import errors
 from threedi_modelchecker.schema import constants
 from threedi_modelchecker.schema import get_schema_version
 from threedi_modelchecker.schema import ModelSchema
+from threedi_modelchecker.threedi_model.views import ALL_VIEWS
 from unittest import mock
 
 import pytest
@@ -86,12 +87,12 @@ def test_validate_schema(threedi_db):
     """Validate a correct schema version"""
     schema = ModelSchema(threedi_db)
     with mock.patch.object(
-        schema, "get_version", return_value=constants.MIN_SCHEMA_VERSION
+        schema, "get_version", return_value=get_schema_version()
     ):
         assert schema.validate_schema()
 
 
-@pytest.mark.parametrize("version", [-1, constants.MIN_SCHEMA_VERSION - 1, None])
+@pytest.mark.parametrize("version", [-1, 205, None])
 def test_validate_schema_missing_migration(threedi_db, version):
     """Validate a too low schema version"""
     schema = ModelSchema(threedi_db)
@@ -112,7 +113,7 @@ def test_validate_schema_too_high_migration(threedi_db, version):
 def test_full_upgrade_empty(in_memory_sqlite):
     """Upgrade an empty database to the latest version"""
     schema = ModelSchema(in_memory_sqlite)
-    schema.upgrade(backup=False)
+    schema.upgrade(backup=False, set_views=False)
     assert schema.get_version() == get_schema_version()
     assert in_memory_sqlite.get_engine().has_table("v2_connection_nodes")
 
@@ -120,15 +121,15 @@ def test_full_upgrade_empty(in_memory_sqlite):
 def test_full_upgrade_with_preexisting_version(south_latest_sqlite):
     """Upgrade an empty database to the latest version"""
     schema = ModelSchema(south_latest_sqlite)
-    schema.upgrade(backup=False)
+    schema.upgrade(backup=False, set_views=False)
     assert schema.get_version() == get_schema_version()
     assert south_latest_sqlite.get_engine().has_table("v2_connection_nodes")
 
 
 def test_full_upgrade_oldest(oldest_sqlite):
-    """Upgrade an empty database to the latest version"""
+    """Upgrade a legacy database to the latest version"""
     schema = ModelSchema(oldest_sqlite)
-    schema.upgrade(backup=False)
+    schema.upgrade(backup=False, set_views=False)
     assert schema.get_version() == get_schema_version()
     assert oldest_sqlite.get_engine().has_table("v2_connection_nodes")
 
@@ -140,7 +141,7 @@ def test_upgrade_south_not_latest_errors(in_memory_sqlite):
         schema, "get_version", return_value=constants.LATEST_SOUTH_MIGRATION_ID - 1
     ):
         with pytest.raises(errors.MigrationMissingError):
-            schema.upgrade(backup=False)
+            schema.upgrade(backup=False, set_views=False)
 
 
 def test_upgrade_with_backup(threedi_db):
@@ -152,7 +153,7 @@ def test_upgrade_with_backup(threedi_db):
         "threedi_modelchecker.schema._upgrade_database", side_effect=RuntimeError
     ) as upgrade, mock.patch.object(schema, "get_version", return_value=199):
         with pytest.raises(RuntimeError):
-            schema.upgrade(backup=True)
+            schema.upgrade(backup=True, set_views=False)
 
     (db,), kwargs = upgrade.call_args
     assert db is not threedi_db
@@ -165,7 +166,20 @@ def test_upgrade_without_backup(threedi_db):
         "threedi_modelchecker.schema._upgrade_database", side_effect=RuntimeError
     ) as upgrade, mock.patch.object(schema, "get_version", return_value=199):
         with pytest.raises(RuntimeError):
-            schema.upgrade(backup=False)
+            schema.upgrade(backup=False, set_views=False)
 
     (db,), kwargs = upgrade.call_args
     assert db is threedi_db
+
+
+def test_set_views(oldest_sqlite):
+    """Make sure that the views are regenerated
+    """
+    schema = ModelSchema(oldest_sqlite)
+    schema.upgrade(backup=False, set_views=True)
+    assert schema.get_version() == get_schema_version()
+
+    # Test all views
+    with oldest_sqlite.session_scope() as session:
+        for view_name in ALL_VIEWS:
+            session.execute(f"SELECT * FROM {view_name} LIMIT 1").fetchall()